refactor: move Stats management to utils and clean up exit logging
