+++ actions/inferGrammar.ts
@@ -1,10 +1,9 @@
 import path from "path";
 import type { CLIInferGrammarArguments } from "../cli";
 import fs from 'fs';
-import { calculateComplexity } from "../heuristics/complexity";
 import { loadOpenAIEnvVars, type OpenAIEnv, type OpenAIMessage } from "../llm/utils";
 import type { ANTLRError } from "../syntactic/ErrorListener";
-import { constructPrompt, generateInitalGuess, makeCompletionRequest, Stats, type Grammar, type Snippet, type TestedSnippet } from '../llm/grammar';
+import { constructPrompt, generateInitalGuess, makeCompletionRequest, type Grammar, type Snippet, type TestedSnippet } from '../llm/grammar';
 import { checkGrammar } from "../syntactic/check-grammar";
 import { _createWorkingDirectory, _createTemporaryFile, findAllCodeFiles, loadFile } from './utils/io';
 import { compressMessages, countTokens } from '../llm/compress-messages';
@@ -12,8 +11,16 @@ import { loadSnippetsByComplexity } from './utils/snippets';
 import type { Result } from '../result';
 import { GrammarGenerator, GrammarVerifier } from '../llm/grammarPipeline';
 import { runInferenceLoop, type InferenceOptions } from '../llm/autoCreationLoop';
+import { ExitAndLogStats as _ExitAndLogStats } from './utils';
 
 const temporaryFileDirectoryRecords = new Set<string>();
+
+function ExitAndLogStats(exitCode: number = 0) {
+    for (const tempPath of temporaryFileDirectoryRecords) {
+        fs.rmSync(tempPath, { recursive: true, force: true });
+    }
+    _ExitAndLogStats(exitCode);
+}
 
 const createWorkingDirectory = () => _createWorkingDirectory('grammar');
 
@@ -225,31 +232,6 @@ export async function repairGrammar(
     // If we reach here, we exhausted the allowed attempts without fully fixing the grammar.
     console.warn(`\n[Repair] Max attempts (${maxRetries}) reached; returning last tested result.`);
     return currentTestedSnippets;
-}
-
-function ExitAndLogStats(exitCode: number = 0) {
-    // Remove temporary file directory
-    for (const tempPath of temporaryFileDirectoryRecords) {
-        fs.rmSync(tempPath, { recursive: true, force: true });
-    }
-
-    console.log("\n[Stats]");
-    console.log(`Generated ${Stats.totalRequests} requests, and completed ${Stats.totalCompletedRequests} requests.`);
-    if (Stats.cachedInputTokens === 0) {
-        console.log(`    Input tokens: ${Stats.inputTokens}, Output tokens: ${Stats.outputTokens}`);
-    } else {
-        const NonCachedTokens = Stats.inputTokens - Stats.cachedInputTokens;
-        console.log(`    Cached input tokens: ${Stats.cachedInputTokens}, Non-cached input tokens: ${NonCachedTokens}, Output tokens: ${Stats.outputTokens}`);
-    }
-
-    console.log(`    ${Stats.totalTokens} tokens (${Stats.avgTokensPerRequest} avg tokens per request)`);
-    if (Stats.score.size > 0) console.log("\n[Model scores]");
-    Array.from(Stats.score.entries())
-        .sort(([, scoreA], [, scoreB]) => scoreB - scoreA)
-        .forEach(([modelName, score]) => {
-            console.log(`    - ${modelName}: ${score}`);
-        });
-    process.exit(exitCode);
 }
 
 export async function buildFirstIntermediateSolution(openaiEnv: OpenAIEnv,
